% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_devices}
\alias{check_devices}
\title{Check PyTorch device availability}
\usage{
check_devices(env_path = NULL)
}
\arguments{
\item{env_path}{Path to virtual environment containing PyTorch and ciss-vae. Defaults to NULL.}
}
\value{
Vector of strings for available devices.
}
\description{
This function prints the available devices (cpu, cuda, mps) detected by PyTorch. If your mps/cuda device is not shown, check your PyTorch installation.
}
\examples{
\donttest{
try(
check_devices()
)}
}
