% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{mock_surv}
\alias{mock_surv}
\title{Example survival data for demo of imputable_matrix}
\format{
A dataframe:
\describe{
\item{mock_surv}{A mock survival dataset}
}
}
\source{
Imported from \code{data_raw/mock_survival.csv}
}
\usage{
mock_surv
}
\description{
A sample survival dataset
}
\examples{
data(mock_surv)
}
\keyword{datasets}
